+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to
 ### Added
 
 - Retry with exponential backoff function to error handling
+- `mfaEnabled` property added to `onelogin_user`, with `undefined` value, for
+  future development
 
 ## 2.2.13 - 2021-09-21